Make feature:home module
