+++ .idea/gradle.xml
@@ -16,6 +16,8 @@
             <option value="$PROJECT_DIR$/data/local" />
             <option value="$PROJECT_DIR$/data/remote" />
             <option value="$PROJECT_DIR$/data/repository" />
+            <option value="$PROJECT_DIR$/feature" />
+            <option value="$PROJECT_DIR$/feature/home" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
