+++ app/src/main/java/com/tfandkusu/template/view/main/AppContent.kt
@@ -1,14 +1,10 @@
 package com.tfandkusu.template.view.main
 
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.tfandkusu.template.ui.theme.AppTemplateTheme
+import com.tfandkusu.template.view.home.HomeScreen
 
 private const val HOME_PATH = "home"
 
@@ -17,26 +13,5 @@ fun AppContent() {
     val navController = rememberNavController()
     NavHost(navController = navController, startDestination = HOME_PATH) {
         composable(HOME_PATH) { HomeScreen() }
-    }
-}
-
-@Composable
-fun HomeScreen() {
-    // A surface container using the 'background' color from the theme
-    Surface(color = MaterialTheme.colors.background) {
-        Greeting("Android")
-    }
-}
-
-@Composable
-fun Greeting(name: String) {
-    Text(text = "Hello $name!")
-}
-
-@Preview(showBackground = true)
-@Composable
-fun DefaultPreview() {
-    AppTemplateTheme {
-        Greeting("Android")
     }
 }
