+++ feature/home/src/main/java/com/tfandkusu/template/view/home/HomeScreen.kt
@@ -1,0 +1,19 @@
+package com.tfandkusu.template.view.home
+
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun HomeScreen() {
+    // A surface container using the 'background' color from the theme
+    Surface(color = MaterialTheme.colors.background) {
+        Greeting("Android")
+    }
+}
+
+@Composable
+fun Greeting(name: String) {
+    Text(text = "Hello $name!")
+}
