+++ settings.gradle
@@ -12,3 +12,4 @@ include ':common'
 include ':data:remote'
 include ':data:local'
 include ':data:repository'
+include ':feature:home'
